backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.2.0
